requirements.txt
async-timeout==4.0.3
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
flask==3.0.3
flask-wtf==1.2.1
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mongo==0.2.0
pymongo==4.7.2
python-dotenv==1.0.1
redis==5.0.4
werkzeug==3.0.3
wtforms==3.1.2
zipp==3.18.2
